이미지 개선중

+++ JavaGUiGame/src/Enemy/WeakEnemy.java
@@ -8,22 +8,22 @@ import status_basis.RectangleHitbox;
 
 public class WeakEnemy extends abstractEnemy {
     private double hp = setHp();
-    private double x = 200; // 변경: int -> double
-    private double y = 200; // 변경: int -> double
+    private int x = 200; // 변경: int -> double
+    private int y = 200; // 변경: int -> double
     private final int SIZE = 50;
     private final double MOVE_AMOUNT = 3; // 변경: int -> double
     private abstractHitbox hitbox;
     private BufferedImage enemyImage;
     private WeakEnemyImage enemyImg;
     
-    private double playerX; // 변경: int -> double
-    private double playerY; // 변경: int -> double
+    private int playerX; // 변경: int -> double
+    private int playerY; // 변경: int -> double
 
     public double getenHp() {
     	return hp;
     }
     
-    public WeakEnemy(double playerX, double playerY) { // 변경: int -> double
+    public WeakEnemy(int playerX, int playerY) { // 변경: int -> double
         setOpaque(false);
         hitbox = new RectangleHitbox((int) x, (int) y, SIZE, SIZE);
         enemyImg = new WeakEnemyImage();
@@ -33,7 +33,7 @@ public class WeakEnemy extends abstractEnemy {
         this.playerY = playerY;
     }
     
-    public void setPlayerPosition(double playerX, double playerY) { // 변경: int -> double
+    public void setPlayerPosition(int playerX, int playerY) { // 변경: int -> double
         this.playerX = playerX;
         this.playerY = playerY;
     }
