+++ JavaGUiGame/src/abstracts/PlayerPositionProvider.java
@@ -1,6 +1,6 @@
 package abstracts;
 
 public interface PlayerPositionProvider {
-    double getPlayerX();
-    double getPlayerY();
+    int getPlayerX();
+    int getPlayerY();
 }
